minor optimization add and bug fix

+++ backend/controllers/products.controller.js
@@ -49,10 +49,13 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const {id} = req.params
     
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(401).json({ success: false, message: "Product not found"})
+    }
     try {
         await Product.findByIdAndDelete(id)
         res.status(200).json({ success: true, message: "Product deleted"})
     } catch(error) {
-        res.status(404).json({ success: false, message: "Product not found"})
+        res.status(500).json({ success: false, message: "server error"})
     }
 }
